+++ screens/TimerScreen.js
@@ -3,8 +3,9 @@ import * as Font from 'expo-font';
 import { StyleSheet, Text, View } from 'react-native';
 import Loader from '../components/Loader';
 import StartRoundButton from '../components/StartRoundButton';
-import Timer from '../components/Timer.js'
+import {Timer} from '../components/Timer.js'
 import { widthPercentageToDP as wp, heightPercentageToDP as hp } from 'react-native-responsive-screen';
+import { Card } from '../components/Card';
 
 export default class TimerScreen extends React.Component {
 
@@ -59,56 +60,11 @@ export default class TimerScreen extends React.Component {
      })
    }
 
-   action = flag => {
-     flag ? this.start() : this.stop()
+   toggle() {
+        !this.state.started ? this.start() : this.stop()
    }
 
    render() {
-     const styles = StyleSheet.create({
-        container: {
-            backgroundColor: '#F0F0F0',
-            flex: 1
-        },
-
-        secondFloor: {
-            marginTop: hp('10%'),
-            width: wp('75%'),
-            height: wp('120%'),
-            shadowColor: "#000",
-            shadowOffset: {width: 0, height: 4},
-            shadowOpacity: 0.37,
-            shadowRadius: 6.49,
-            elevation: 8,
-            borderRadius: 15,
-            borderColor: "#E1E1E1",
-            backgroundColor: "#fff",
-            alignItems: 'center',
-        },
-
-        timerCircle: {
-            marginTop: wp('6%'),
-            width: wp('64%'),
-            height: wp('64%'),
-            borderTopLeftRadius: wp('32%'),
-            borderTopRightRadius: wp('32%'),
-            borderBottomLeftRadius: wp('32%'),
-            borderBottomRightRadius: wp('32%'),
-            borderColor: this.state.circleBorder,
-            borderWidth: 2,
-            backgroundColor: "#fff",
-            shadowColor: "#000",
-            shadowOffset: {width: 0, height: 2},
-            shadowOpacity: 0.2,
-            shadowRadius: 2.49,
-            elevation: 3,
-            justifyContent: 'center',
-            textAlign: 'center',
-            alignItems: 'center',
-        }
-
-     })
-
-
        if (!this.state.fontLoaded) {
            return (
              <View style={{flex: 1, justifyContent: 'center', alignItems: 'center', textAlign: 'center'}}>
@@ -117,26 +73,24 @@ export default class TimerScreen extends React.Component {
            )
        } else {
            return (
-              <View style={{flex: 1, backgroundColor: "#F0F0F0", alignItems: 'center'}}>
-                 <View style={styles.secondFloor}>
-                         <View style={styles.timerCircle}>
-                            <Timer mil = {this.state.diff}/>
-                          </View>
-                          <View style={{marginTop: wp('25%')}}>
-                          <View style={{flexDirection: 'row'}}>
-                              <Text style={{color: "#000", fontSize: wp('5%')}}>2 Referee</Text>
-                              <Text style={{color: "#666666", fontSize: wp('5%'), fontFamily: 'roboto-mono', marginLeft: wp('14%')}}>01:18:03</Text>
-                          </View>
-                          <View style={{flexDirection: 'row', marginTop: wp('3%')}}>
-                              <Text style={{color: "#000", fontSize: wp('5%')}}>3 Referee</Text>
-                              <Text style={{color: "#666666", fontFamily: 'roboto-mono', fontSize: wp('5%'), marginLeft: wp('14%')}}>01:20:09</Text>
-                          </View>
-                      </View>
-                  </View>
-                  <View style={{marginTop: hp('4.3%'), justifyContent: 'center', alignItems: 'center'}}>
-                      <StartRoundButton action={(flag) => {this.action(flag)}} />
-                  </View>
-              </View>
+              <Card
+              circle = { ()=> <Timer mil = {this.state.diff}/> }
+              text = { ()=> 
+                <View>
+                <Text style={{color: "#000", fontSize: wp('5%')}}>
+                    2 Referee: 
+                    <Text style={{color: "#666666", fontSize: wp('5%'), fontFamily: 'roboto-mono', marginLeft: wp('4%')}}> 01:18:03</Text>
+                </Text>
+                <Text style={{color: "#000", fontSize: wp('5%')}}>
+                    3 Referee: 
+                    <Text style={{color: "#666666", fontSize: wp('5%'), fontFamily: 'roboto-mono', marginLeft: wp('4%')}}> 01:18:03</Text>
+                </Text>
+                </View>
+              }
+              button = { !this.state.started ? "Start" : "Stop" }
+              onPress = {()=>this.toggle()}
+              color = { this.state.circleBorder }
+              />
            );
        }
    }
